% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice0dice.R
\name{slice0dice}
\alias{slice0dice}
\title{Split a string based on given indices}
\usage{
slice0dice(x, split)
}
\arguments{
\item{x}{A string}

\item{split}{A numerical vector indicating which indices of x to be split on}
}
\value{
A character vector
}
\description{
Split a string based on given indices
}
\examples{
x<-"Happy birthday to me"
split<-c(2,9)
splice0dice(x, split)

}
